Big changes
Character now keeps track of slopes alongside floors, to give programmers more information on the terrain. Impulsing now does deactivates when the floor is past the snap distance instead of when the vertical speed is going downwards.

CharacterJump can now be deflected by the floor normal, and jumps can be done on slopes (and reset on slopes)