Broke up CharacterMovement, and refactored basic_animator
The role of character movement was too muddled. It was mainly for rotation and controlling the movement_vector. However, controlling the movement vector is such a trivial task that it wasn't necessary to be in the script. So the remaining function of rotation was broken up into separate scripts: FaceMovement and CopyBasis

Also, BasicAnimator, and any other solution to animation would be too specific to each set of character animations. So I removed the export vars and made it specific to Gobot